html/semantics/scripting-1/the-script-element/scriptexecute-events/031.html

<!DOCTYPE html>
<html>
  <title>No onafterscriptexecute DOM attribute</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    events = [];
    var t = async_test();

    var script = document.createElement("script");
    script.textContent = "events.push('next script ran'); setTimeout(t.step_func(function() {assert_array_equals(events, ['next script ran']); t.done()}), 500)"

   script.onafterscriptexecute =  t.step_func(function() {assert_unreached()});
   document.body.appendChild(script);
 </script>
